backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pandas==2.2.2
chromadb==0.5.0
google-generativeai==0.7.2
langgraph==0.1.19
langchain-core==0.2.9
langchain-google-genai==1.0.6
pydantic==2.7.4
python-dotenv==1.0.1
aiofiles==23.2.1
